RecorderThread: Write duration to FLAC STREAMINFO metadata block
MediaCodec's FLAC encoder does not write the final duration to the
STREAMINFO metadata block. The streaming nature of the API makes that
impossible. Instead, when the recording completes, BCR will find the
STREAMINFO block and update the total frame count field manually.

This fixes seeking in audio players that don't try to guess the length
of the audio file.

Fixes: #19
Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>