RecorderThread: Record to internal storage first
This commit changes RecorderThread's behavior to first record to the
default/fallback output directory and then move to the user-specified
target directory when the recording is complete.

This avoids issues where writing to the SAF-provided file descriptor
sometimes has several hundren millisecond latencies, causing
AudioRecord's internal buffer to overrun. On some devices, this happens
even when the SAF directory is on internal storage. The root cause of
why the SAF-provided file descriptor is slow is unknown. Perhaps the
FUSE filesystem at /sdcard does some additional checks on write with SAF
paths compared to the Context.getExternalFilesDir() path.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>