SettingsActivity: Prompt for optional contacts permission on first enable
The contacts permission is used to add the contact name to the output
filenames. The permission prompt will only be shown the first time call
recording is enabled. If the user chooses to not allow the permission,
but wants to allow it later, they will have to do so via the system
settings.

Fixes: #78
Fixes: #80

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>