RecorderThread: Use non-blocking reads from AudioRecord
On Samsung devices, when a call ends, AudioRecord.read() blocks until
another call becomes active, which prevents recordings from stopping at
the correct time. This behavior does not happen in AOSP.

To work around this, use non-blocking reads with a sleep interval of the
AudioRecord minimum buffer size. This gives the recording loop a chance
to check if recording has been cancelled.

Issue: #143

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>