Send notification when a call recording completes
This commit adds a new notification that fires when a call recording
completes successfully. It's associated with a new notification channel
so it can be disabled without affecting other notifications.

The notification has two action buttons: open and share. Both of these
do as their name implies and do not dismiss the notification. Clicking
on the notification text behaves like open, except the notification is
dismissed.

The open and share options require a custom ContentProvider because it's
not possible to grant the target app permissions to a SAF URI. BCR's
RecorderProvider is very minimal and just wraps SAF's provider for the
openFile, getType, and query operations.

Also, manual action from the user is required for the new open and share
options to work. Previously, BCR only requested persistent write
permissions to the output directory. This makes it impossible for
RecorderProvider to access the recording that was just created. This
commit updates the output directory permission logic to also request
persistent read permissions, but the user will have to reset the output
directory to the default and then select the desired directory again.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>