Explicitly vibrate when vibration is enabled for a notification channel
Android's vibration setting seemingly has no effect when the user is in
a phone call, which is always the case when a BCR notification is shown.
With this commit, BCR will explicitly vibrate the device using a pattern
that mimics what Android does by default for notifications.

This commit is inspired by @quyenvsp's PR #167, except it relies on
Android notification channels' builtin vibration setting instead of
adding a new setting to BCR.

Related PR: #167

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>