RecorderThread: Fix silent crash when debug mode is enabled
The output filename variable was not initialized at the point that
startLogcat() was called, resulting in an uninitialized property
exception being thrown. This bug was introduced in commit
588d3f69b71a2b0691b0a0e29b397d32b7a8dd3f.

This commit also fixes startLogcat() accessing the filename property
without acquiring the filenameLock lock. In practice, this did not cause
issues, but better to be correct.

Discovered in issue #195.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>