Add new option for starting the recording in the paused state
This commit adds a new "initially paused" option to BCR's UI. If it's
disabled, then BCR works as it always has: calls are recorded as soon as
they enter the connected state. If the option is enabled, then the
recording starts in the paused state. The user can choose to resume via
BCR's persistent notification. If the recording is never resumed, then
the empty output file is deleted and no success notification will be
shown.

This new option is disabled by default.

Issue: #198

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>