magisk: Try to install to a non-overlayfs location
Magisk currently has a known issue where its mirror mount points do not
work correctly with overlayfs. Files from the overlayfs lowerdirs may
no longer be accessible.

On MIUI, this is a problem because /system/etc/permissions is an
overlayfs mount and one of the lowerdirs is
/product/pango/system/etc/permissions. One file underneath that path is
com.nxp.nfc.nq.xml, which if missing, breaks NFC.

Fixes: #242
Fixes: #246

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>