Add support for writing call metadata to an adjacent JSON file
The JSON file contains all of the metadata known to BCR (which roughly
matches the information that can be used in the filename templates).
This feature is disabled by default.

The majority of this commit is splitting the call metadata collection
out of OutputFilenameGenerator. This is something that should've been
done anyway.

The notification delete button logic has also been reworked so that the
metadata file is also deleted. If debug mode is enabled, the log file
will now also be deleted.

Closes: #380

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>