Launch Android Settings to re-enable battery optimization
Currently, BCR will just gray out the `Disable battery optimization`
option after it is enabled because Android does not provide a way for
apps to re-enable optimizations.

This isn't very user-friendly, so this commit updates the option to
never be grayed out, but instead, take the user to the appropriate page
in Android's Settings.

Fixes: #422

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>