Switch to format-specific sample rates
Previously, the sample rate was a global option with choices that are
known to work with every supported format. Most of these choices are not
compatible with some upcoming formats, so switch to format-specific
sample rates. As part of this change, selecting any valid sample rate
supported by Android's encoders is now possible.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>