Switch DirectBootMigrationService to a background service
There seem to be two issues with the foreground service:

* Some Android builds seem to be broken and throw a
  ForegroundServiceDidNotStartInTimeException even though the service
  unconditionally calls startForeground() in onStartCommand(). The root
  cause of this is unknown. Perhaps these devices are exceptionally slow
  during boot?
* The persistent notification does not get dismissed when the file
  migration runs too quickly.

We may be able to work around the second issue by not setting the
FOREGROUND_SERVICE_IMMEDIATE notification behavior flag, but given how
fast the file migration is, let's just switch to a background service.
We're unlikely to hit Android's limitations on background services.

Closes: #574
Closes: #576

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>