Split pure silence warning notification to separate channel
The warning can sometimes be triggered when hanging up before the other
party is able to answer, so allow the user to disable the notification
if desired.

This commit also changes the recorder thread status reporting to be more
explicit and type-safe. All user-facing error messages are now computed
in the UI layers instead of the recorder thread.

Closes: #579

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>