service.sh: Restore SELinux labels on DP data directory during boot
OxygenOS on OnePlus devices seems to create the /data/user_de/0/<id>
directory with the wrong SELinux label initially. However, once the
correct label is set, it persists just fine. Despite it persisting, we
still need to run restorecon during boot instead of during installation
because it requires PackageManager to already be aware of BCR.

Fixes: #582
Fixes: #591
Fixes: #598

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>