Add support for specifying a minimum recording duration
If the recording isn't long enough, it will be discarded at the end of
the call. The duration is computed based on the recording duration
(number of encoded frames), not the call duration, so it excludes
regions where the user explicitly paused the recording and regions where
the call was placed on hold.

Similar to the record rules, this can be overridden during the call from
BCR's notification.

Fixes: #411
Fixes: #604

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>